+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Version 2.1
 
 - Added a new workflow with 2 groups of ControlNet nodes for pose and hand control
+- Added examples
 
 ## 17.12.2023