+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Version 2.1
 - Added body type option
 - Added wrinkles option
 - Added beard option
+- Added tanned skin option
 - Added a new workflow with 2 groups of ControlNet nodes for pose and hand control
 - Added examples